+++ playlet-lib/src/locale/es_ES/translations.ts
@@ -1248,5 +1248,9 @@
         <source>Select keyboard language</source>
         <translation>Seleccionar idioma del teclado</translation>
     </message>
+    <message>
+        <source>Rewind live TV</source>
+        <translation>Retroceder TV en directo</translation>
+    </message>
 </context>
 </TS>