+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Live stream videos for Playlet backend, attempt with Invidious
 - Audio track names when using Invidious
+- Small leak in video player
 
 ## [0.45.0] - 2026-05-25